+++ lib/ooxml_parser/common_parser/common_data/colors/hsl_color.rb
@@ -19,39 +19,6 @@ module OoxmlParser
       @l = lightness
     end
 
-    # Convert rgb color to hsl
-    # @param rgb_color [Color] color to convert
-    # @return [HSLColor] result of convert
-    def self.rgb_to_hsl(rgb_color)
-      hls_color = HSLColor.new
-      red = rgb_color.red.to_f # / 255.0
-      green = rgb_color.green.to_f # / 255.0
-      blue = rgb_color.blue.to_f # / 255.0
-
-      min = [red, green, blue].min.to_f
-      max = [red, green, blue].max.to_f
-
-      delta = (max - min).to_f
-      hls_color.l = (min + max) / 255.0 / 2.0
-      hls_color.alpha_channel = rgb_color.alpha_channel.to_f / 255.0
-
-      unless max == min
-        hls_color.s = delta / (255.0 - (255.0 - max - min).abs)
-
-        hls_color.h = if max == red && green >= blue
-                        60.0 * (green - blue) / delta
-                      elsif max == red && green < blue
-                        60.0 * (green - blue) / delta + 360.0
-                      elsif max == green
-                        60.0 * (blue - red) / delta + 120.0
-                      else
-                        60.0 * (red - green) / delta + 240.0
-                      end
-      end
-
-      hls_color
-    end
-
     # Chroma - The "colorfulness relative to the brightness of a similarly illuminated white".
     # @return [Color] result
     def to_rgb