+++ lib/ooxml_parser/common_parser/common_data/colors/theme_colors.rb
@@ -15,7 +15,7 @@ module OoxmlParser
       # TODO: if no swap performed - incorrect color parsing. But don't know why it needed
       themes_array[0], themes_array[1] = themes_array[1], themes_array[0]
       themes_array[2], themes_array[3] = themes_array[3], themes_array[2]
-      hls = HSLColor.rgb_to_hsl(themes_array[theme].color)
+      hls = themes_array[theme].color.to_hsl
       tint = 0 if tint.nil?
       hls.calculate_rgb_with_tint(tint)
     end